fix(console): keep the launch path off the config write path (CHOO-2228)
Reading what to launch an agent with, and saving its configuration, had
been put in one module. That made the two runtime tests fail to even load:
saving has to reach a remote agent's launch spec, so it depends on the
sidecar machinery — and that machinery asks what to launch with, so the
module imported itself in a circle and dragged the database in with it.

Split the reading side out. It needs none of that: it opens the agent's
working directory and reads a file. The write side keeps the sidecar
dependency and takes the reader's workspace helper, so the cycle is gone
and the launch paths import the lighter of the two.

The runtime tests mock the reader, as they already mock the agent row —
resolving an agent's working directory means a database they do not have.
The two that are about the Codex profile now state the values they assert
on directly, rather than by way of an agent row the production code no
longer reads them from.
