+++ console/apps/switch-console-desktop/src/main/core/sidecar/sidecar-diagnostics.ts
@@ -1,6 +1,6 @@
 import { DEEPLINK_SCHEME } from '@main/app/deeplinks';
 import { readAgentSidecarLog } from '@main/core/agent-runtime/impl/ensure-agent-sidecar';
-import { agentLaunchSpecialization } from '@main/core/agents/agent-config';
+import { agentLaunchSpecialization } from '@main/core/agents/agent-launch-config';
 import { getRemoteAgentLocation } from '@main/core/agents/agent-location';
 import { connectRemoteAgent } from '@main/core/agents/connect-remote-agent';
 import { getAgents } from '@main/core/agents/getAgents';
